stack::l2cap Properly type tL2CAP_LE_RESULT_CODE
Value in spec is 16 bits.

Version 5.4 | Vol 3, Part A 4.26 L2CAP_CREDIT_BASED_CONNECTION_RSP (CODE 0x18) Table 4.17

Bug: 350490212
Test: m .
Flag: EXEMPT, Mechanical Refactor

+++ system/stack/include/l2cdefs.h
@@ -114,7 +114,7 @@
 
 /* Define the LE L2CAP Connection Response Result codes
  */
-typedef enum : uint8_t {
+typedef enum : uint16_t {
   L2CAP_LE_RESULT_CONN_OK = 0,
   L2CAP_LE_RESULT_NO_PSM = 2,
   L2CAP_LE_RESULT_NO_RESOURCES = 4,
